# docs/swagger.yaml
openapi: 3.0.3
info:
    title: Midnight Lace API
    description: >
        API REST para la aplicación *Midnight Lace*.


        **GRUPO 1:**
          - Abeledo, Federico
          - Ghillino, Rocío Belén
          - Novello, Victoria Abril
          - Romero, Mailén Belén


        ---


        ## Clasificación de endpoints


        ### Públicos (sin autenticación)

        Accesibles sin token. Usados para registro, login y datos auxiliares.

        - `POST /auth/registro`


        - `POST /auth/confirmar`

        - `POST /auth/login`

        - `POST /auth/renovar`

        - `GET  /paises`

        - `GET  /paises/{id}`



        ### Protegidos por rol: comprador

        Requieren JWT con rol `comprador`. Todo usuario registrado y admitido los tiene.

        - `GET/PATCH /perfil`


        - `GET/POST  /medios-de-pago`

        - `DELETE    /medios-de-pago/{id}`

        - `GET/POST  /productos` (el POST crea el artículo y otorga rol dueño si no lo tiene)

        - `GET       /productos/{id}`

        - `GET       /subastas`

        - `GET       /subastas/{id}`

        - `GET       /subastas/{id}/catalogo`

        - `GET       /subastas/{id}/item-actual`

        - `GET/POST  /subastas/{idSubasta}/items/{idItem}/pujas`

        - `GET       /mi/subastas`

        - `GET       /mi/pujas`

        - `GET       /mi/compras`

        - `GET       /mi/multas`

        - `POST      /mi/multas/{id}/pagar`

        - `POST      /mi/compras/{id}/pagar`

...

...

...

...



...

...

...

...

...

...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
        pagina:
            name: pagina
            in: query
            description: Número de página (base 1)
            required: false
            schema:
                type: integer
                default: 1
                minimum: 1
        cantidad:
            name: cantidad
            in: query
            description: Cantidad de resultados por página
            required: false
            schema:
                type: integer
                default: 20
                minimum: 1
                maximum: 100
        pathId:
            name: id
            in: path
            required: true
            schema:
                type: integer
    schemas:
...
        MetaPaginacion:
            type: object
            properties:
                pagina:
                    type: integer
                    example: 1
                cantidad:
                    type: integer
                    example: 20
                total:
                    type: integer
                    example: 150
                totalPaginas:
                    type: integer
                    example: 8

        # --------------------------------------------------------
        # ERRORES
        # --------------------------------------------------------
        Error:
            type: object
            properties:
                codigo:
                    type: string
                    description: >
                        Código de error legible por el frontend para tomar decisiones.
                        Códigos posibles en pujas:
                        PUJA_MONTO_INSUFICIENTE — el importe no supera la última oferta + 1% del precio base.
                        PUJA_MONTO_EXCEDIDO — el importe supera la última oferta + 20% del precio base (solo aplica a subastas comun/especial/plata).
                        PUJA_EN_PROCESO — hay una puja anterior aún siendo confirmada, reintentar.
                        PUJA_ITEM_INACTIVO — el ítem ya no está en subasta (se agotó el tiempo o fue vendido).
                        PUJA_CHEQUE_SIN_FONDOS — el montoDisponible del cheque certificado no alcanza.
                    example: PUJA_MONTO_EXCEDIDO
                mensaje:
                    type: string
                    description: Mensaje legible para mostrar al usuario.
                    example: "El importe supera el máximo permitido para esta subasta."
                detalle:
                    type: object
                    description: >
                        Información adicional específica del error. Presente solo cuando
                        aplica. En errores de puja incluye los límites calculados para
                        que el frontend pueda mostrarlos sin hacer otro request.
                    properties:
                        pujaMinima:
                            type: number
                            description: Mínimo válido para la próxima puja.
                            example: 15150.00
                        pujaMaxima:
                            type: number
                            nullable: true
                            description: Máximo válido. Null para subastas oro/platino.
                            example: 17000.00
        ErrorNoAutenticado:
            type: object
            properties:
                codigo:
                    type: string
                    example: NO_AUTENTICADO
                mensaje:
                    type: string
                    example: No autenticado. Iniciá sesión para continuar.
        ErrorSinPermiso:
            type: object
            properties:
                codigo:
                    type: string
                    example: SIN_PERMISO
                mensaje:
                    type: string
                    example: No tenés permisos para realizar esta acción.
        ErrorNoEncontrado:
            type: object
            properties:
                codigo:
                    type: string
                    example: NO_ENCONTRADO
                mensaje:
                    type: string
                    example: El recurso solicitado no existe.
        ErrorValidacion:
            type: object
            properties:
                codigo:
                    type: string
                    example: ERROR_VALIDACION
                mensaje:
                    type: string
                    example: Datos inválidos o incompletos.

        # --------------------------------------------------------
        # AUTH
        # --------------------------------------------------------
...
        Producto:
            type: object
            properties:
                identificador:
                    type: integer
                titulo:
                    type: string
                    maxLength: 200
                estado:
                    type: string
                    enum: [nuevo, usado]
                    description: Condición física del producto.
                fechaIngreso:
                    type: string
                    format: date
                disponible:
                    type: string
                    enum: [si, no]
                descripcionCatalogo:
                    type: string
                descripcionCompleta:
                    type: string
                declaracionPropiedad:
                    type: boolean
                idDuenio:
                    type: integer
                precioBase:
                    type: number
                    description: Precio base fijado por el dueño al cargar el producto.
                nroPoliza:
                    type: string
                    description: Número de póliza de seguro asociada al producto
                estadoProducto:
                    type: string
                    enum:
                        [
                            pendiente,
                            asignado,
                            rechazado,
                            pendiente_confirmacion,
                            en_subasta,
                            vendido,
                        ]
                    description: >
                        pendiente — recién creado, esperando verificación automática.
                        asignado — verificación aprobada, en el pool de un subastador.
                        rechazado — verificación fallida. Estado terminal, se devuelve con cargo.
                        pendiente_confirmacion — agregado a un catálogo con precio y comisión, esperando confirmación del dueño. Si el dueño rechaza, vuelve a asignado.
                        en_subasta — dueño aceptó condiciones, siendo subastado.
                        vendido — vendido exitosamente.
                deposito:
                    $ref: "#/components/schemas/Deposito"
                fotos:
                    type: array
                    items:
                        $ref: "#/components/schemas/FotoProducto"
                detalleArtistico:
                    $ref: "#/components/schemas/DetalleArtistico"
                componentes:
                    type: array
                    items:
                        $ref: "#/components/schemas/ComponenteProducto"
        FotoProducto:
            type: object
            properties:
                identificador:
                    type: integer
                url:
                    type: string
                orden:
                    type: integer
        DetalleArtistico:
            type: object
            properties:
                artista:
                    type: string
                fechaObra:
                    type: string
                    format: date
                historia:
                    type: string
        ComponenteProducto:
            type: object
            properties:
                identificador:
                    type: integer
                descripcion:
                    type: string
                cantidad:
                    type: integer
        SolicitudCrearProducto:
            type: object
            required:
                [
                    titulo,
                    estado,
                    descripcionCompleta,
                    declaracionPropiedad,
                    precioBase,
                    foto1,
                    foto2,
                    foto3,
                    foto4,
                    foto5,
                    foto6,
                ]
            properties:
                titulo:
                    type: string
                    minLength: 1
                    maxLength: 200
                estado:
                    type: string
                    enum: [nuevo, usado]
                descripcionCompleta:
                    type: string
                declaracionPropiedad:
                    type: boolean
                    description: >
                        El dueño declara que el bien le pertenece y no tiene
                        impedimento legal para venderlo. Debe ser true para continuar.
                precioBase:
                    type: number
                    minimum: 0.01
                    description: Precio base de venta fijado por el dueño.
                foto1:
                    type: string
                    format: binary
                    description: Foto obligatoria 1 de 6.
                foto2:
                    type: string
                    format: binary
                    description: Foto obligatoria 2 de 6.
                foto3:
                    type: string
                    format: binary
                    description: Foto obligatoria 3 de 6.
                foto4:
                    type: string
                    format: binary
                    description: Foto obligatoria 4 de 6.
                foto5:
                    type: string
                    format: binary
                    description: Foto obligatoria 5 de 6.
                foto6:
                    type: string
                    format: binary
                    description: Foto obligatoria 6 de 6.
                foto7:
                    type: string
                    format: binary
                    description: Foto adicional opcional.
                foto8:
                    type: string
                    format: binary
                    description: Foto adicional opcional.
                detallesArtisticos:
                    type: string
                    description: >
                        JSON serializado con artista, fechaObra e historia.
                        Se envía como string dentro del multipart porque el body es form-data.
                    example: '{"artista":"Vivienne Westwood","fechaObra":"1995-01-01","historia":"..."}'
                componentes:
                    type: string
                    description: >
                        JSON serializado con array de componentes (descripcion, cantidad).
                        Se envía como string dentro del multipart.
                    example: '[{"descripcion":"Sombrero","cantidad":1},{"descripcion":"Guantes","cantidad":2}]'

        # --------------------------------------------------------
        # SEGURO
        # --------------------------------------------------------
        Seguro:
            type: object
            properties:
                nroPoliza:
                    type: string
                compania:
                    type: string
                polizaCombinada:
                    type: string
                    enum: [si, no]
                importe:
                    type: number

        # --------------------------------------------------------
        # CUENTA DE COBRO (dueño)
        # --------------------------------------------------------
...
        Deposito:
            type: object
            properties:
                identificador:
                    type: integer
                nombre:
                    type: string
                    example: "Depósito Central"
                direccion:
                    type: string
                    example: "Av. Corrientes 1234, CABA"

        # --------------------------------------------------------
        # SUBASTA
        # --------------------------------------------------------
...
paths:
...
    /productos:
        get:
            tags: [Productos]
            summary: Listar productos del dueño autenticado
            description: >
                **Rol requerido: dueño.**
                Devuelve solo los productos del usuario autenticado.
            security:
                - bearerAuth: []
            parameters:
                - $ref: "#/components/parameters/pagina"
                - $ref: "#/components/parameters/cantidad"
                - name: estado
                  in: query
                  required: false
                  schema:
                      type: string
                      enum:
                          [
                              pendiente,
                              asignado,
                              rechazado,
                              pendiente_confirmacion,
                              en_subasta,
                              vendido,
                          ]
            responses:
                "200":
                    description: Lista de productos del dueño
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    datos:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/Producto"
                                    meta:
                                        $ref: "#/components/schemas/MetaPaginacion"
                "401":
                    description: No autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoAutenticado"
                "403":
                    description: El usuario no tiene rol de dueño
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorSinPermiso"

        post:
            tags: [Productos]
            summary: Cargar un artículo para subastar
            description: >
                **Rol requerido: comprador.**
                El usuario carga un artículo para que la empresa lo evalúe.
                Si es la primera vez que carga un artículo, el backend hace un INSERT
                en la tabla `duenios` usando el mismo `identificador` de `personas`,
                verificando primero que no exista ya un registro (INSERT condicional).
                A partir de ese momento el JWT del usuario incluye el rol `duenio`
                además de `comprador`. Si ya era dueño, no se crea nada nuevo.
                La declaración de propiedad es obligatoria (debe ser true).
                La inspección es automática y aleatoria en esta implementación:
                si pasa, el estadoProducto queda en "asignado" y se asigna un
                subastador al azar. Si no pasa, queda en "rechazado".
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            $ref: "#/components/schemas/SolicitudCrearProducto"
            responses:
                "201":
                    description: Artículo cargado. Estado inicial según resultado de verificación automática.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Producto"
                "400":
                    description: Datos inválidos (ej. menos de 6 fotos, declaración de propiedad en false)
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorValidacion"
                "401":
                    description: No autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoAutenticado"
    /productos/{id}:
        get:
            tags: [Productos]
            summary: Ver detalle de un producto
            description: >
                **Rol requerido: dueño.**
                Solo el dueño del producto puede verlo.
            security:
                - bearerAuth: []
            parameters:
                - $ref: "#/components/parameters/pathId"
            responses:
                "200":
                    description: Detalle del producto
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Producto"
                "401":
                    description: No autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoAutenticado"
                "403":
                    description: El producto no pertenece al dueño autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorSinPermiso"
                "404":
                    description: Producto no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoEncontrado"
    /productos/{id}/seguro:
        get:
            tags: [Productos]
            summary: Ver póliza de seguro de un producto
            description: >
                **Rol requerido: dueño.**
                El dueño puede ver la póliza contratada por la empresa para su artículo.
                El cliente puede contactar a la compañía aseguradora para ampliar el monto de la póliza.
            security:
                - bearerAuth: []
            parameters:
                - $ref: "#/components/parameters/pathId"
            responses:
                "200":
                    description: Datos de la póliza de seguro
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Seguro"
                "401":
                    description: No autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoAutenticado"
                "403":
                    description: El producto no pertenece al dueño autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorSinPermiso"
                "404":
                    description: Producto sin seguro asignado todavía
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoEncontrado"
    /productos/{id}/aceptar-condiciones:
        patch:
            tags: [Productos]
            summary: Aceptar o rechazar las condiciones propuestas por el subastador
            description: >
                **Rol requerido: dueño.**
                Disponible solo cuando el producto está en estado "pendiente_confirmacion".
                Si el dueño acepta, el `estadoProducto` pasa a "en_subasta".
                Si el dueño rechaza, el `estadoProducto` vuelve a "asignado", se crea una
                notificación `devolucion_producto` con los gastos de devolución y el
                producto es removido del catálogo.
            security:
                - bearerAuth: []
            parameters:
                - $ref: "#/components/parameters/pathId"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [acepta]
                            properties:
                                acepta:
                                    type: boolean
            responses:
                "200":
                    description: Decisión registrada. Producto actualizado.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Producto"
                "400":
                    description: Datos inválidos
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorValidacion"
                "401":
                    description: No autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoAutenticado"
                "403":
                    description: El producto no pertenece al dueño autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorSinPermiso"
                "404":
                    description: Producto no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoEncontrado"
                "409":
                    description: El producto no está en estado pendiente_confirmacion
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"

    # ============================================================
    # CUENTAS DE COBRO — Dueño
    # ============================================================
...
    /subastador/productos:
        get:
            tags: [Subastador]
            summary: Listar el pool de productos del subastador autenticado
            description: >
                **Rol requerido: subastador.**
                Devuelve los productos asignados al subastador autenticado.
                Los productos en estado "asignado" están disponibles para agregar a un catálogo.
            security:
                - bearerAuth: []
            parameters:
                - $ref: "#/components/parameters/pagina"
                - $ref: "#/components/parameters/cantidad"
                - name: estado
                  in: query
                  required: false
                  schema:
                      type: string
                      enum:
                          [
                              asignado,
                              pendiente_confirmacion,
                              en_subasta,
                              vendido,
                          ]
            responses:
                "200":
                    description: Lista de productos del pool
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    datos:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/Producto"
                                    meta:
                                        $ref: "#/components/schemas/MetaPaginacion"
                "401":
                    description: No autenticado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorNoAutenticado"
                "403":
                    description: Requiere rol subastador
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorSinPermiso"

    # ============================================================
    # EMPLEADO (admin)
    # ============================================================
...